+++ lib/kb.rb
@@ -23,7 +23,7 @@ require 'kb/inflections'
 
 require 'kb/cache'
 require 'kb/client_resolver'
-require 'kb/error'
+require 'kb/errors'
 require 'kb/client'
 
 require 'kb/concerns'