+++ lib/kb/errors.rb
@@ -1,0 +1,6 @@
+require 'kb/errors/error'
+require 'kb/errors/client_error'
+
+require 'kb/errors/conflict_error'
+require 'kb/errors/unprocessable_entity_error'
+require 'kb/errors/resource_not_found'