+++ lib/kb/errors/client_error.rb
@@ -1,0 +1,9 @@
+module KB
+  class ClientError < Error
+    def message
+      JSON.parse(body)['message']
+    rescue StandardError
+      body
+    end
+  end
+end