+++ lib/kb/models/concerns/updatable.rb
@@ -11,7 +11,7 @@ module KB
         api_response = kb_client.update(key, attributes)
         attributes_from_response(api_response)
       rescue Faraday::Error => e
-        raise KB::Error.new(e.response[:status], e.response[:body], e)
+        raise KB::Error.from_faraday(e)
       end
     end
   end