+++ lib/kb/models/concerns/upsertable.rb
@@ -10,7 +10,7 @@ module KB
       def upsert(attributes)
         from_api kb_client.upsert(attributes)
       rescue Faraday::Error => e
-        raise KB::Error.new(e.response[:status], e.response[:body], e)
+        raise KB::Error.from_faraday(e)
       end
     end
   end