+++ lib/kb/models/pet_parent.rb
@@ -29,8 +29,8 @@ module KB
 
     private_class_method :attributes_from_response
 
-    STRING_FIELDS = %i[ key partner_name first_name last_name prefix_phone_number
-                        phone_number email country address zip_code nif].freeze
+    STRING_FIELDS = %i[key partner_name first_name last_name prefix_phone_number
+                       phone_number email country address zip_code nif].freeze
     DATE_FIELDS = %i[birth_date deleted_at].freeze
     FIELDS = [*STRING_FIELDS, *DATE_FIELDS].freeze