+++ spec/fake/models/pet_parent_spec.rb
@@ -13,7 +13,7 @@ RSpec.describe KB::PetParent do
     subject(:upserted_pet_parent) { described_class.upsert(attributes.merge(first_name: first_name)) }
 
     let(:attributes) do
-      { phone_number: '683123123', prefix_phone_number: '+34' }
+      { phone_number: '683123123', prefix_phone_number: '+34', email: 'foo@example.com' }
     end
     let(:first_name) { 'JMJ' }
 
@@ -21,6 +21,9 @@ RSpec.describe KB::PetParent do
       before { existing_pet_parent }
 
       let(:existing_pet_parent) { described_class.create(attributes) }
+      let(:phone_attributes) do
+        { phone_number: '680000000', prefix_phone_number: '+34' }
+      end
 
       it 'return the existing PetParent' do
         expect(upserted_pet_parent.key).to eq(existing_pet_parent[:key])
@@ -28,6 +31,36 @@ RSpec.describe KB::PetParent do
 
       it 'updates the entity with the provided properties' do
         expect(upserted_pet_parent.first_name).to eq(first_name)
+      end
+
+      context 'when updating with a non existent phone number' do
+        subject(:upserted_pet_parent) { described_class.upsert(attributes.merge(phone_attributes)) }
+
+        it 'updates the entity with the provided phone number' do
+          expect(upserted_pet_parent.phone_number).to eq(phone_attributes[:phone_number])
+        end
+      end
+
+      context 'when updating with an existent phone number' do
+        subject(:upserted_pet_parent) do
+          described_class.upsert(attributes.merge(phone_number: phone_attributes[:phone_number]))
+        end
+
+        before { other_pet_parent }
+
+        let(:other_pet_parent) { described_class.create(phone_attributes) }
+
+        it 'raises an error' do
+          expect { upserted_pet_parent }.to raise_error(KB::ConflictError)
+        end
+      end
+
+      context 'when updating the email' do
+        subject(:upserted_pet_parent) { described_class.upsert(attributes.merge(email: 'bar@example.com')) }
+
+        it 'raises an error' do
+          expect { upserted_pet_parent }.to raise_error(KB::UnprocessableEntityError)
+        end
       end
     end